requirements.txt
streamlit==1.32.0
python-docx==1.1.0
google-generativeai>=0.8.0
